#286043 refactored max
